requirements.txt
google-cloud-translate==2.0.1
pandas
spacy
scikit-learn
numpy==1.23
pillow
matplotlib
nltk
tqdm
torchinfo
torchtext==0.13.0
slackclient
wandb
sentence-transformers
openpyxl
xlrd
seaborn
markupsafe==2.0.1
scikit-image

# for detectron2
docutils==0.16
# https://github.com/sphinx-doc/sphinx/commit/7acd3ada3f38076af7b2b5c9f3b60bb9c2587a3d
sphinx==3.2.0
recommonmark==0.6.0
sphinx_rtd_theme
# Dependencies here are only those required by import
termcolor
yacs
tabulate
cloudpickle
Pillow
future
fvcore
omegaconf>=2.1.0.dev24
hydra-core>=1.1.0.dev5
scipy
timm==v0.5.4
fairscale
opencv-python
pycocotools
mss
ftfy
lvis
tensorboard

wandb

git+https://github.com/cocodataset/panopticapi.git